+++ src/cgame/sp/cg_sound.cpp
@@ -44,9 +44,9 @@ int numSoundScriptSounds = 0;
 return a hash value for the filename
 ================
 */
-static long generateHashValue( const char *fname ) {
+static int32_t generateHashValue( const char *fname ) {
 	int i;
-	long hash;
+	int32_t hash;
 	char letter;
 
 	hash = 0;
@@ -59,7 +59,7 @@ static long generateHashValue( const char *fname ) {
 		if ( letter == '\\' ) {
 			letter = '/';                   // damn path names
 		}
-		hash += (long)( letter ) * ( i + 119 );
+		hash += (int32_t)( letter ) * ( i + 119 );
 		i++;
 	}
 	hash &= ( FILE_HASH_SIZE - 1 );
@@ -75,7 +75,7 @@ CG_SoundScriptPrecache
 */
 int CG_SoundScriptPrecache( const char *name ) {
 	soundScriptSound_t *scriptSound;
-	long hash;
+	int32_t hash;
 	char *s;
 	soundScript_t   *sound;
 	byte buf[1024];
@@ -178,7 +178,7 @@ CG_SoundPlaySoundScript
 ==============
 */
 qboolean CG_SoundPlaySoundScript( const char *name, vec3_t org, int entnum ) {
-	long hash;
+	int32_t hash;
 	char *s;
 	soundScript_t   *sound;
 
@@ -237,7 +237,7 @@ static void CG_SoundParseSounds( char *filename, const char *buffer ) {
 	char* token;
 	const char** text;
 	int s;
-	long hash;
+	int32_t hash;
 	soundScript_t sound;                // the current sound being read
 	soundScriptSound_t  *scriptSound;
 	qboolean inSound, wantSoundName;
