+++ src/client/cl_net_chan.cpp
@@ -98,7 +98,7 @@ CL_Netchan_Decode
 ==============
 */
 static void CL_Netchan_Decode( msg_t *msg ) {
-	long reliableAcknowledge, i, index;
+	int32_t reliableAcknowledge, i, index;
 	byte key, *string;
 	int srdc, sbit, soob;
