+++ src/game/botlib.h
@@ -247,7 +247,7 @@ typedef struct botlib_import_s
 	int ( *FS_Read )( void *buffer, int len, fileHandle_t f );
 	int ( *FS_Write )( const void *buffer, int len, fileHandle_t f );
 	void ( *FS_FCloseFile )( fileHandle_t f );
-	int ( *FS_Seek )( fileHandle_t f, long offset, int origin );
+	int ( *FS_Seek )( fileHandle_t f, int32_t offset, int origin );
 	//debug visualisation stuff
 	int ( *DebugLineCreate )( void );
 	void ( *DebugLineDelete )( int line );
@@ -458,10 +458,10 @@ typedef struct ai_export_s
 	void ( *BotEnterChat )( int chatstate, int client, int sendto );
 	void ( *BotGetChatMessage )( int chatstate, char *buf, int size );
 	int ( *StringContains )( char *str1, char *str2, int casesensitive );
-	int ( *BotFindMatch )( char *str, struct bot_match_s *match, unsigned long int context );
+	int ( *BotFindMatch )( char *str, struct bot_match_s *match, uint32_t context );
 	void ( *BotMatchVariable )( struct bot_match_s *match, int variable, char *buf, int size );
 	void ( *UnifyWhiteSpaces )( char *string );
-	void ( *BotReplaceSynonyms )( char *string, unsigned long int context );
+	void ( *BotReplaceSynonyms )( char *string, uint32_t context );
 	int ( *BotLoadChatFile )( int chatstate, char *chatfile, char *chatname );
 	void ( *BotSetChatGender )( int chatstate, int gender );
 	void ( *BotSetChatName )( int chatstate, char *name );
