+++ src/renderer/tr_backend.cpp
@@ -373,8 +373,8 @@ void GL_TexEnv( int env ) {
 ** This routine is responsible for setting the most commonly changed state
 ** in Q3.
 */
-void GL_State( unsigned long stateBits ) {
-	unsigned long diff = stateBits ^ glState.glStateBits;
+void GL_State( uint32_t stateBits ) {
+	uint32_t diff = stateBits ^ glState.glStateBits;
 
 	if ( !diff ) {
 		return;
@@ -2512,7 +2512,7 @@ const void  *RB_SwapBuffers( const void *data ) {
 	// counting up the number of increments that have happened
 	if ( r_measureOverdraw->integer ) {
 		int i;
-		long sum = 0;
+		int32_t sum = 0;
 		unsigned char *stencilReadback;
 
 		stencilReadback = static_cast<byte*> (ri.Hunk_AllocateTempMemory( glConfig.vidWidth * glConfig.vidHeight ));
